pyproject.toml
[project]
name = "starteragent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "livekit-agents[silero,turn-detector]~=1.2",
    "livekit-plugins-assemblyai~=1.2",
    "livekit-plugins-cartesia~=1.2",
    "livekit-plugins-deepgram~=1.2",
    "livekit-plugins-openai~=1.2",
    "livekit-plugins-noise-cancellation~=0.2",
    "python-dotenv>=1.1.1",
    "fastapi>=0.112.0",
    "uvicorn[standard]>=0.30.0",
    "supabase>=2.4.0",
    "pydantic>=2.7",
    "pydantic-settings>=2.0",
    "sqlalchemy>=2.0",
    "alembic>=1.13",
]
